requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2025.4
Django==4.2.20
django-crispy-forms==2.4
django-widget-tweaks==1.5.0
pillow==11.1.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.2
